test nexus push

+++ src/main/java/com/esprit/examen/services/StockServiceImpl.java
@@ -6,6 +6,8 @@ import java.util.Date;
 import java.util.List;
 import java.util.logging.Logger;
 
+import com.esprit.examen.Exceptions.CatchException;
+import com.esprit.examen.dto.StockDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.esprit.examen.entities.Stock;
@@ -24,11 +26,11 @@ public class StockServiceImpl implements IStockService {
 
 
 	@Override
-	public List<Stock> retrieveAllStocks() {
+	public List<StockDto> retrieveAllStocks() {
 		// récuperer la date à l'instant t1
 		log.info("Begin Processing - Retreive all Stock Method ! ");
-		List<Stock> stocks = (List<Stock>) stockRepository.findAll();
-		stocks.forEach(stock -> {log.info("Stock : " +stock);} );
+		List<StockDto> stocks =  StockDto.fromEntityList( stockRepository.findAll());
+		stocks.forEach(stock -> log.info("Stock : " +stock) );
 		log.info("End Processing - Retreive all Stock Method ! \"");
 		// récuperer la date à l'instant t2
 		// temps execution = t2 - t1
@@ -36,38 +38,43 @@ public class StockServiceImpl implements IStockService {
 	}
 
 	@Override
-	public Stock addStock(Stock s) {
+	public StockDto addStock(StockDto s) {
 		// récuperer la date à l'instant t1
 		log.info("Begin Processing addStock");
-		return stockRepository.save(s);
+		stockRepository.save(StockDto.toEntity(s));
+		log.info("End Processing addStock");
+		return s ;
 		
 	}
 
 	@Override
 	public void deleteStock(Long stockId) {
 		log.info("Begin Processing deleteStock");
+		if(stockId==null) throw new CatchException(" id is null ");
 		stockRepository.deleteById(stockId);
 		log.info("End Processing deleteStock");
 
 	}
 
 	@Override
-	public Stock updateStock(Long id ,  Stock s) {
+	public StockDto updateStock(Long id ,  StockDto s) {
 		log.info("Begin Processing updateStock");
 		s.setIdStock(id);
-		return stockRepository.save(s);
+		stockRepository.save(StockDto.toEntity(s));
+		log.info("End Processing updateStock");
+		return s ;
 	}
 
 	@Override
-	public Stock retrieveStock(Long stockId) {
+	public StockDto retrieveStock(Long stockId) {
 		long start = System.currentTimeMillis();
 		log.info("Begin Processing -  retrieveStock");
 		Stock stock = stockRepository.findById(stockId).orElse(null);
 		log.info("End  processing  _ retrieveStock");
-		 long elapsedTime = System.currentTimeMillis() - start;
+		long elapsedTime = System.currentTimeMillis() - start;
 		log.info("Method execution time: " + elapsedTime + " milliseconds.");
 
-		return stock;
+		return StockDto.fromEntity(stock);
 	}
 
 	@Override
@@ -76,19 +83,19 @@ public class StockServiceImpl implements IStockService {
 		SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss.SSS");
 		Date now =  new Date();
 		String msgDate = sdf.format(now);
-		String finalMessage = "";
+		StringBuilder finalMessage =  new StringBuilder();
 		String newLine = System.getProperty("line.separator");
-		List<Stock> stocksEnRouge = (List<Stock>) stockRepository.retrieveStatusStock();
+		List<Stock> stocksEnRouge =  stockRepository.retrieveStatusStock();
 		for (int i = 0; i < stocksEnRouge.size(); i++) {
-			finalMessage = newLine + finalMessage + msgDate + newLine + ": le stock "
+			finalMessage.append(newLine + finalMessage + msgDate + newLine + ": le stock "
 					+ stocksEnRouge.get(i).getLibelleStock() + " a une quantité de " + stocksEnRouge.get(i).getQte()
 					+ " inférieur à la quantité minimale a ne pas dépasser de " + stocksEnRouge.get(i).getQteMin()
-					+ newLine;
+					+ newLine);
 
 		}
-		log.info(finalMessage);
+		log.info(finalMessage.toString());
 		log.info("End Processing - retrieveStatusStock ()");
-		return finalMessage;
+		return finalMessage.toString();
 	}
 
 }
